+++ client/src/types/dave.ts
@@ -1,0 +1,26 @@
+export enum DaveState {
+  STANDING,
+  RUNNING,
+  JUMPING_DOWN,
+  JUMPING_UP,
+  FALLING,
+  SHOOTING,
+  RECHARGING,
+}
+
+export enum DaveMove {
+  LEFT,
+  RIGHT,
+  NONE,
+}
+
+export enum DaveLook {
+  LEFT,
+  RIGHT,
+}
+
+export enum DaveShoot {
+  UP,
+  CENTER,
+  DOWN,
+}
